requirements.txt
pdf2image
pillow
google-generativeai